Working API

+++ FiguresApi/Contracts/Coordinates.cs
@@ -2,8 +2,8 @@
 {
     public class Coordinates
     {
-        public decimal X { get; set; }
+        public double X { get; set; }
 
-        public decimal Y { get; set; }
+        public double Y { get; set; }
     }
 }
